requirements.txt
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
opencv-python>=4.8.0
albumentations>=1.3.1
tqdm>=4.65.0
Pillow>=9.5.0
pandas>=2.0.0
scikit-learn>=1.3.0
gradio>=4.0.0
matplotlib>=3.7.0
realitydefender>=0.1.0
python-dotenv>=1.0.0
